test: moves setup for unit tests in /tests/unit and replaces imports
