+++ src/components/MultiStepForm/Steps/__tests__/Geolocation.spec.tsx
@@ -8,7 +8,7 @@ import userEvent from "@testing-library/user-event";
 
 import Geolocation from "../Geolocation";
 import MultiStepFormWrapper from "../../MultiStepFormWrapper";
-import createFetchResponse from "../../../../lib/__mocks__/fetch";
+import createFetchResponse from "@/tests/unit/fetch";
 import type { Values } from "@/types";
 import { normalizeCity } from "@/utils";
