+++ src/tests/unit/db.ts
@@ -1,4 +1,3 @@
-import { vi } from "vitest";
 import { mockDeep, DeepMockProxy } from "vitest-mock-extended";
 import { PrismaClient } from "@prisma/client";
 import { db } from "@/lib";
